Simplify path handling in tutorials/export
